Refactor flower code into seperate component
