+++ src/components/App.jsx
@@ -1,7 +1,6 @@
-import Constants from '../constants';
+import Flower from './Flower.jsx';
 import React from 'react';
 import ReactCanvas from 'react-canvas';
-import { randElement, randBetween } from '../util';
 
 
 export default class App extends React.Component {
@@ -15,20 +14,15 @@ export default class App extends React.Component {
     this._windowDidResize();
     this._setupCanvas();
 
-    const randX = randBetween(this.state.width);
-    const randY = randBetween(this.state.height);
-
-    this._createCanvasCircle(this._context, {
-      xPos: randX,
-      yPos: randY,
-      radius: Constants.flowerRadius,
-      color: randElement(Constants.flowerColors)
-    });
+    setInterval(() => {
+      const flower = new Flower(this._context, this.state);
+      flower.draw();
+    }, 500);
   }
 
   _windowDidResize() {
     window.onresize = event => {
-      this.setState(this.getWindowDimensions());
+      this.setState(this._getWindowDimensions());
     };
   }
 
@@ -42,18 +36,6 @@ export default class App extends React.Component {
   _setupCanvas() {
     this._canvas = React.findDOMNode(this);
     this._context = this._canvas.getContext('2d');
-  }
-
-  _createCanvasCircle(ctx, opts = {
-                                    xPos, 
-                                    yPos, 
-                                    radius: 1,
-                                    color: '#000000'
-                                  }) {
-    const path = new Path2D();
-    path.arc(opts.xPos, opts.yPos, opts.radius, 0, Math.PI * 2, true);
-    ctx.fillStyle = opts.color;
-    ctx.fill(path);
   }
 
   render() {
